django2/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cssselect2==0.4.1
defusedxml==0.6.0
deploy-to-heroku==0.0.1
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-heroku-connect==0.23.0
django-stdimage==5.1.1
django3-easy-pdf==0.1.2.dev1
Djheroku==0.6.1
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPrompt==0.1.1
GitPython==3.1.12
gunicorn==20.0.4
heroku==0.1.4
heroku-git-fs==1.0.0
herokutils==0.0.1
html5lib==1.1
idna==2.10
joblib==1.0.0
mysql==0.0.2
mysqlclient==2.0.2
nltk==3.5
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.0
progressbar2==3.53.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==1.0.2
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==1.5
python-utils==2.4.0
pytz==2020.4
regex==2020.11.13
reportlab==3.5.59
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
smmap==3.0.5
soupsieve==2.1
sqlparse==0.4.1
static3==0.7.0
threadpoolctl==2.1.0
tinycss2==1.1.0
tqdm==4.56.0
urllib3==1.26.3
virtualenv==20.2.2
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
